2.0.12

+++ Pod/Classes/Services/KWS/KWSCheckRegistered.m
@@ -23,7 +23,7 @@
     return GET;
 }
 
-- (void) successWithStatus:(int)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
+- (void) successWithStatus:(NSInteger)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
     if (!success) {
         _checkRegistered(false, false);
     } else {
