+++ Pod/Classes/Services/KWS/KWSInviteUser.m
@@ -29,7 +29,7 @@
     };
 }
 
-- (void) successWithStatus:(int)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
+- (void) successWithStatus:(NSInteger)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
     _invited (success && (status == 200 || status == 204));
 }
