+++ Pod/Classes/Services/KWS/KWSRequestPermission.m
@@ -41,7 +41,7 @@
     };
 }
 
-- (void) successWithStatus:(int)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
+- (void) successWithStatus:(NSInteger)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
     if (!success) {
         _requested(false, false);
     } else {
