+++ Pod/Classes/Services/KWS/KWSTriggerEvent.m
@@ -37,7 +37,7 @@
     return evtDict;
 }
 
-- (void) successWithStatus:(int)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
+- (void) successWithStatus:(NSInteger)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
     _triggered (success && (status == 200 || status == 204));
 }
