+++ Pod/Classes/Services/KWS/KWSUnregisterToken.m
@@ -30,7 +30,7 @@
     };
 }
 
-- (void) successWithStatus:(int)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
+- (void) successWithStatus:(NSInteger)status andPayload:(NSString *)payload andSuccess:(BOOL)success {
     _unregisteredToken (success && (status == 200 || status == 204));
 }
